MAJOR-CHANGE! Migration of integer id's to Long id's (including fk's etc.)

+++ plugins/org.projectforge.plugins.datatransfer/src/main/kotlin/org/projectforge/plugins/datatransfer/DataTransferAuditDO.kt
@@ -71,7 +71,7 @@ open class DataTransferAuditDO {
   @get:Id
   @get:GeneratedValue
   @get:Column(name = "pk")
-  open var id: Int? = null
+  open var id: Long? = null
 
   @get:Column
   open var timestamp: Date? = null
@@ -80,7 +80,7 @@ open class DataTransferAuditDO {
    * No constraint (area might be deleted in the mean time).
    */
   @get:Column(name = "area_fk", nullable = false)
-  open var areaId: Int? = null
+  open var areaId: Long? = null
 
   @get:ManyToOne(fetch = FetchType.LAZY)
   @get:JoinColumn(name = "by_user_fk")
