+++ plugins/org.projectforge.plugins.extendemployeedata/src/main/java/org/projectforge/plugins/eed/service/EmployeeSalaryImportService.java
@@ -80,7 +80,7 @@ public class EmployeeSalaryImportService
     for (final ImportedElement<EmployeeSalaryDO> el : sheet.getElements()) {
       final EmployeeSalaryDO employeeSalary = el.getValue();
       if (employeeSalary != null) {
-        final Integer id = employeeSalary.getId();
+        final Long id = employeeSalary.getId();
         final EmployeeSalaryDO dbEmployeeSalary = (id != null) ? employeeSalaryService.selectByPk(id) : null;
         if (dbEmployeeSalary != null) {
           el.setOldValue(dbEmployeeSalary);
